requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
requests>=2.26.0,<3.0.0
pytest>=6.2.5,<7.0.0
python-dotenv>=0.19.0,<1.0.0
pymongo>=4.3.3,<5.0.0
dnspython>=2.3.0,<3.0.0
